+++ projects/mind-mapp/app/src/utils/index.ts
@@ -8,6 +8,7 @@ export * from './id';
 export * from './importValidation';
 export * from './sampleMap';
 export * from './selectionBounds';
+export * from './selectionText';
 export * from './shortcuts';
 export * from './storage';
 export * from './uiPrefs';